==> 06_merging_counts.Rmd <==
---
title: "06 — Merging counts data"
output: html_document
---

# Merging

```{r, message = F}
# load packages
require(readr)
require(dplyr)
require(stringr)
require(tidyr)
require(purrr)
require(ggplot2)

# set working directory
wd <- "C:/Users/Daniel/Google Drive/Courses/S&DS 315/S&DS 315 Final Project/Code"
setwd(wd)

# read clicks data
clicks <- read_csv("../Data/05_clicks.csv")

# we had 6 experiments, labeled A through F (letters is built-in vector)
experiments <- letters[1:6]

# Function to read in, for each experiment, the csv that contains
#   the treatment assignment and whether the text was sent,
#   then merge in whether the link was clicked
merge_clicks <- function(exp_label) {
	read_csv(paste0("../data/04_assignments_sent_", exp_label, ".csv"),
			 col_types = cols(
	             .default = col_guess(),
	             Experiment = col_character() # otherwise, it will convert "F" to FALSE
	         )) %>%
		mutate(slashtag = str_replace(Shortlink, "go.teatexts.me/", "")) %>%
		left_join(clicks, by = "slashtag") %>%
		mutate(Clicked = as.numeric(clicks > 0)) %>%
		select(Phone, Experiment, Treatment, Success, Clicked)
}

# merge, store in list, ten combine into one long df
merged <- lapply(experiments, merge_clicks) %>%
	bind_rows()
```

# Click rate by experiment

```{r}
roundPct <- function(val) {
	(val * 100) %>% round(1) %>% paste0("%")
}

# extrafont::font_import(pattern = "Roboto")
windowsFonts(RobotoCondensed = windowsFont("Roboto Condensed"))
windowsFonts(RobotoCondensedBold = windowsFont("Roboto Condensed Bold"))
```

For the notebook:

```{r}
merged %>%
	group_by(Experiment, Treatment) %>%
	# there are 4 NAs because for some reason, a link was never generated, so they were never sent a text
	summarize(PctClicked = mean(Clicked, na.rm = TRUE)) %>%
	mutate(Treatment = as.factor(if_else(Treatment == 0, "No", "Yes"))) %>%
	ggplot(aes(x = Treatment, y = PctClicked, fill = Treatment)) +
	geom_bar(stat = "identity") +
	geom_text(aes(label = roundPct(PctClicked)), vjust = -1, color = "black", size = 4, family = "RobotoCondensed") +
	facet_grid(cols = vars(Experiment)) +
	labs(y = "Percent clicked",
		 title = "Summary of click rates",
		 subtitle = "By treatment group and experiment\n") +
	scale_y_continuous(limits = c(0, 0.12),
					   labels = scales::percent_format(accuracy = 1L),
					   breaks = seq(0, 0.12, by = 0.03)) +
	scale_fill_manual(values = c("#195896", "#ee6c4d")) +
	theme_minimal() +
	theme(text = element_text(family = "RobotoCondensed"),
		  plot.title = element_text(family = "RobotoCondensedBold", size = 18),
		  plot.subtitle = element_text(size = 14),
		  panel.grid.major = element_blank(),
		  panel.grid.minor = element_blank(),
		  axis.title = element_text(size = 14),
		  axis.title.y = element_text(margin = margin(r = 10)),
		  axis.title.x = element_text(margin = margin(t = 12)),
		  axis.text = element_text(size = 11),
		  legend.position = "none",
		  strip.text = element_text(size = 12, family = "RobotoCondensed"))
```

For the JPEG file:

```{r}
jpeg("summary.jpg", width = 12, height = 6, units = "in", res = 300)

merged %>%
	group_by(Experiment, Treatment) %>%
	# there are 4 NAs because for some reason, a link was never generated, so they were never sent a text
	summarize(PctClicked = mean(Clicked, na.rm = TRUE)) %>%
	mutate(Treatment = as.factor(if_else(Treatment == 0, "No", "Yes"))) %>%
	ggplot(aes(x = Treatment, y = PctClicked, fill = Treatment)) +
	geom_bar(stat = "identity") +
	geom_text(aes(label = roundPct(PctClicked)), vjust = -1, color = "black", size = 6, family = "RobotoCondensed") +
	facet_grid(cols = vars(Experiment)) +
	labs(y = "Percent clicked",
		 title = "Summary of click rates",
		 subtitle = "By treatment group and experiment\n\n") +
	scale_y_continuous(limits = c(0, 0.12),
					   labels = scales::percent_format(accuracy = 1L),
					   breaks = seq(0, 0.12, by = 0.03)) +
	scale_fill_manual(values = c("#195896", "#ee6c4d")) +
	theme_minimal() +
	theme(text = element_text(family = "RobotoCondensed"),
		  plot.title = element_text(family = "RobotoCondensedBold", size = 24),
		  plot.subtitle = element_text(size = 16),
		  axis.title = element_text(size = 18),
		  axis.title.y = element_text(margin = margin(r = 12)),
		  axis.title.x = element_text(margin = margin(t = 15)),
		  axis.text = element_text(size = 16),
		  panel.grid.major = element_blank(),
		  panel.grid.minor = element_blank(),
		  # legend.text = element_text(size = 11),
		  # legend.title = element_text(size = 14),
		  # legend.spacing.y = unit(0.25, "cm"),
		  # legend.key.size = unit(0.75, "cm"),
		  legend.position = "none",
		  strip.text = element_text(size = 20, family = "RobotoCondensedBold"))

dev.off()
```

# ATE by experiment

Preparation (creating the covariates dataframe):

```{r}
covariates <- read_csv("../data/02_responses_cleaned.csv") %>%
	rename(NumberOfTeas = `Number of teas`) %>%
	mutate(Graduation = factor(Graduation))

significance <- merged %>%
	left_join(covariates) %>%
	nest(data = -Experiment) %>%
	mutate(model = map(data, ~ lm(Clicked ~ Treatment, data = .)),
		   model_grad = map(data, ~ lm(Clicked ~ Treatment * Graduation, data = .)),
		   model_num = map(data, ~ lm(Clicked ~ Treatment * NumberOfTeas, data = .)),
		   ate = map_dbl(model, ~ .$coefficients["Treatment"]),
		   sd = map_dbl(model, ~ summary(.)$coefficients["Treatment", 2]),
		   ci_low = ate - 1.96 * sd,
		   ci_high = ate + 1.96 * sd)
```

For the notebook:

```{r}
ggplot(significance, aes(x = Experiment, y = ate)) +
	geom_bar(stat = "identity", fill = "#195896") +
	geom_errorbar(aes(ymin = ci_low, ymax = ci_high), width = 0.2, color = "#ee6c4d", size = 1) +
	geom_point(color = "#ee6c4d", size = 3) +
	geom_text(aes(label = roundPct(ate), vjust = ifelse(ate > 0, -1.15, 2)), hjust = 1.3, color = "black", size = 4, family = "RobotoCondensed") +
	geom_hline(yintercept = 0, color = "#195896") +
	labs(y = "Average treatment effect",
		 title = "Average treatment effects with 95% CI",
		 subtitle = "Estimated using linear regression, without covariates or blocking") +
	theme_minimal() +
	theme(text = element_text(family = "RobotoCondensed"),
		  plot.title = element_text(family = "RobotoCondensedBold", size = 18),
		  plot.subtitle = element_text(size = 14),
		  axis.title = element_text(size = 14),
		  axis.title.y = element_text(margin = margin(r = 10)),
		  axis.title.x = element_text(margin = margin(t = 12)),
		  axis.text = element_text(size = 11),
		  panel.grid.major = element_blank(),
		  panel.grid.minor = element_blank())
```

For the JPEG file:

```{r}
jpeg("ate.jpg", width = 12, height = 6, units = "in", res = 300)

ggplot(significance, aes(x = Experiment, y = ate)) +
	geom_bar(stat = "identity", fill = "#195896") +
	geom_errorbar(aes(ymin = ci_low, ymax = ci_high), width = 0.2, color = "#ee6c4d", size = 1) +
	geom_point(color = "#ee6c4d", size = 5) +
	geom_text(aes(label = roundPct(ate), vjust = ifelse(ate > 0, -1.15, 2)), hjust = 1.3, color = "black", size = 6, family = "RobotoCondensed") +
	geom_hline(yintercept = 0, color = "#195896") +
	labs(y = "Average treatment effect",
		 title = "Average treatment effects with 95% CI",
		 subtitle = "Estimated using linear regression, without covariates or blocking") +
	theme_minimal() +
	theme(text = element_text(family = "RobotoCondensed"),
		  plot.title = element_text(family = "RobotoCondensedBold", size = 24),
		  plot.subtitle = element_text(size = 16),
		  axis.title = element_text(size = 18),
		  axis.title.y = element_text(margin = margin(r = 12)),
		  axis.title.x = element_text(margin = margin(t = 25)),
		  axis.text = element_text(size = 16),
		  panel.grid.major = element_blank(),
		  panel.grid.minor = element_blank())

dev.off()
```

Some notes on getting the fonts to work:

- I couldn't get `hrbrthemes` to work properly using anything other than Arial Narrow and Roboto Condensed
- `extrafont::font_import(pattern = "Roboto")` will search by Pattern
- Make sure fonts are installed to the default Windows font folder
- For more help, see
	- https://github.com/hrbrmstr/hrbrthemes/issues/28
	- https://stackoverflow.com/questions/14733732/cant-change-fonts-in-ggplot-geom-text
	- https://stackoverflow.com/questions/34522732/changing-fonts-in-ggplot2
- See above for syntax on saving images in high-res

```{r}
jpeg("ate_hte_grad.jpg", width = 6, height = 6, units = "in", res = 300)

merged %>%
	left_join(covariates) %>%
	mutate(Treatment = as.factor(if_else(Treatment == 0, "No", "Yes"))) %>%
	group_by(Experiment, Treatment, Graduation) %>%
	summarize(PctClicked = mean(Clicked, na.rm = TRUE)) %>%
	ggplot(aes(x = Treatment, y = PctClicked, color = Graduation)) +
	geom_line(aes(group = Graduation), size = 1, stat = "identity") +
	geom_point(aes(group = Graduation), size = 2, stat = "identity") +
	facet_wrap(~Experiment) +
	labs(y = "Percent clicked",
		 title = "Click rate by graduation year",
		 subtitle = "No hetergeneous treatment effect by graduation year\n") +
	theme_minimal() +
	scale_y_continuous(limits = c(0, 0.2),
					   labels = scales::percent_format(accuracy = 1L),
					   breaks = seq(0, 0.2, by = 0.1)) +
	scale_color_manual(values = c("#195896", "#70b2d8", "#ee6c4d", "#60c659")) +
	theme(text = element_text(family = "RobotoCondensed"),
		  plot.title = element_text(family = "RobotoCondensedBold", size = 24),
		  plot.subtitle = element_text(size = 16),
		  axis.title = element_text(size = 18),
		  axis.title.y = element_text(margin = margin(r = 20)),
		  axis.title.x = element_text(margin = margin(t = 20)),
		  axis.text = element_text(size = 16),
		  legend.text = element_text(size = 11),
		  legend.title = element_text(size = 14),
		  legend.spacing.y = unit(0.25, "cm"),
		  legend.key.size = unit(0.75, "cm"),
		  strip.text = element_text(size = 16, family = "RobotoCondensed"),
		  panel.grid.major.x = element_blank())

dev.off()
```

```{r}
jpeg("ate_hte_num.jpg", width = 6, height = 6, units = "in", res = 300)

merged %>%
	left_join(covariates) %>%
	mutate(Treatment = as.factor(if_else(Treatment == 0, "No", "Yes"))) %>%
	group_by(Experiment, Treatment, NumberOfTeas) %>%
	summarize(PctClicked = mean(Clicked, na.rm = TRUE)) %>%
	ggplot(aes(x = Treatment, y = PctClicked, color = NumberOfTeas)) +
	geom_line(aes(group = NumberOfTeas), size = 1, stat = "identity") +
	geom_point(aes(group = NumberOfTeas), size = 2, stat = "identity") +
	facet_wrap(~Experiment) +
	labs(y = "Percent clicked",
		 title = "Click rate by # of previous teas",
		 subtitle = "No hetergeneous treatment effect by # of prev. teas\n",
		 color = "# prev. teas") +
	theme_minimal() +
	scale_y_continuous(limits = c(0, 0.2),
					   labels = scales::percent_format(accuracy = 1L),
					   breaks = seq(0, 0.2, by = 0.1)) +
	scale_color_manual(values = c("#195896", "#70b2d8", "#ee6c4d", "#60c659")) +
	theme(text = element_text(family = "RobotoCondensed"),
		  plot.title = element_text(family = "RobotoCondensedBold", size = 24),
		  plot.subtitle = element_text(size = 16),
		  axis.title = element_text(size = 18),
		  axis.title.y = element_text(margin = margin(r = 20)),
		  axis.title.x = element_text(margin = margin(t = 20)),
		  axis.text = element_text(size = 16),
		  legend.text = element_text(size = 11),
		  legend.title = element_text(size = 14),
		  legend.spacing.y = unit(0.25, "cm"),
		  legend.key.size = unit(0.75, "cm"),
		  strip.text = element_text(size = 16, family = "RobotoCondensed"),
		  panel.grid.major.x = element_blank())

dev.off()

# merged %>%
# 	left_join(covariates) %>%
# 	mutate(Treatment = as.factor(if_else(Treatment == 0, "No", "Yes"))) %>%
# 	group_by(Treatment, Graduation) %>%
# 	summarize(PctClicked = mean(Clicked, na.rm = TRUE)) %>%
# 	ggplot(aes(x = Treatment, y = PctClicked, color = Graduation)) +
# 	geom_line(aes(group = Graduation), stat = "identity") +
# 	theme_minimal() +
# 	theme(text = element_text(family = "RobotoCondensed"),
# 		  plot.title = element_text(family = "RobotoCondensedBold", size = 18),
# 		  plot.subtitle = element_text(size = 12)) +
# 	labs(y = "Percent clicked",
# 		 title = "Interaction plot of click rate by graduation year",
# 		 subtitle = "No hetergeneous treatment effect by graduation year")
```
